requirements.txt
#
####### requirements.txt #######
#

#
####### Eve              #######
####### 0.9 - latest   #######
#
eve == 0.9

#
####### Flask CORS       #######
####### Latest 3.0.7     #######

flask-cors == 3.0.7

#https://pypi.org/project/dependency-injector/
dependency-injector

#https://interface.readthedocs.io/en/latest/
python-interface

psycopg2

Flask-SQLAlchemy

typing

python-dotenv
